requirements.txt
biopython>=1.79
psutil>=5.8.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
numpy>=1.21.0
requests>=2.27.0
tqdm>=4.62.0
fastapi>=0.95.0
uvicorn[standard]>=0.20.0